修复`Axis`的`Time`时间轴无法显示大于2038的年份的问题

+++ Runtime/Component/Axis/AxisHandler.cs
@@ -364,7 +364,7 @@ namespace XCharts
             {
                 var lastCount = axis.context.labelValueList.Count;
                 axis.context.tickValue = DateTimeUtil.UpdateTimeAxisDateTimeList(axis.context.labelValueList,
-                    (int)axis.context.minValue, (int)axis.context.maxValue, axis.splitNumber);
+                    axis.context.minValue, axis.context.maxValue, axis.splitNumber);
 
                 if (axis.context.labelValueList.Count != lastCount)
                     axis.SetAllDirty();
